move files of item widgets to product folder
